# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Trademark Monitoring Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /monitoring/keywords:
    get:
      operationId: listMonitoringKeywords
      tags: [monitoring]
      summary: List monitoring keywords
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SearchParam"
        - name: country
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Keywords
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedMonitoringKeywords"
    post:
      operationId: addMonitoringKeyword
      tags: [monitoring]
      summary: Add monitoring keyword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddKeywordBody"
      responses:
        "200":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /monitoring/keywords/{id}:
    delete:
      operationId: deleteMonitoringKeyword
      tags: [monitoring]
      summary: Remove keyword
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /monitoring/keywords/{id}/requeue-scopes:
    post:
      operationId: requeueMonitoringKeywordScopes
      tags: [monitoring]
      summary: Remove existing scopes for keyword and enqueue fresh ones
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scopes re-queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /monitoring/results:
    get:
      operationId: listMonitoringResults
      tags: [monitoring]
      summary: List monitoring results
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SearchParam"
        - name: country
          in: query
          description: Source country(s); comma-separated for OR (PHP multi-select)
          schema:
            type: string
        - name: journalDate
          in: query
          schema:
            type: string
        - name: journalDates
          in: query
          description: Comma-separated journal dates (OR)
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: integer
        - name: clientIds
          in: query
          description: Comma-separated client ids (OR)
          schema:
            type: string
        - name: keywordId
          in: query
          schema:
            type: integer
        - name: keywords
          in: query
          description: Comma-separated watch keywords (exact match on monitoring_keywords.keyword)
          schema:
            type: string
        - name: classFilter
          in: query
          description: Comma-separated Nice class numbers; filters conflict_class (PHP)
          schema:
            type: string
        - name: conflictClasses
          in: query
          schema:
            type: string
        - name: conflict
          in: query
          description: Substring match on conflict tm applied for
          schema:
            type: string
        - name: conflictAppno
          in: query
          schema:
            type: string
        - name: scores
          in: query
          description: Comma-separated 100,90,80,70 bucket filters
          schema:
            type: string
        - name: sortBy
          in: query
          description: Column to sort by (PHP TM Watch default journalDate desc)
          schema:
            type: string
            enum:
              - journalDate
              - sourceKeyword
              - country
              - conflict
              - appno
              - class
              - client
              - score
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedMonitoringResults"
  /monitoring/results/{id}/favorite:
    post:
      operationId: toggleMonitoringResultFavorite
      tags: [monitoring]
      summary: Toggle favorite star (PHP ViewResults favorite)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: New favorite state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitoringResultFavoriteToggle"
        "403":
          description: Forbidden
        "404":
          description: Not found
  /monitoring/results/export:
    post:
      operationId: exportMonitoringResults
      tags: [monitoring]
      summary: Export monitoring results
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportBody"
      responses:
        "200":
          description: Export file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResponse"
  /monitoring/import:
    post:
      operationId: importMonitoringKeywords
      tags: [monitoring]
      summary: Import keywords from file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                tmname:
                  type: string
                class:
                  type: string
                country:
                  type: string
                client:
                  type: string
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResponse"
...
components:
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    SearchParam:
      name: search
      in: query
      schema:
        type: string
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
      required: [status, message]
...
    MonitoringKeyword:
      type: object
      properties:
        id:
          type: integer
        keyword:
          type: string
        country:
          type: string
        class:
          type: string
        userId:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        userName:
          type: string
        userEmail:
          type: string
        createdAt:
          type: string
      required: [id, keyword, country, class]
    PaginatedMonitoringKeywords:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MonitoringKeyword"
        total:
          type: integer
        page:
          type: integer
        perPage:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, perPage, totalPages]
    AddKeywordBody:
      type: object
      properties:
        keyword:
          type: string
        country:
          type: string
        class:
          type: string
        clientId:
          type: integer
      required: [keyword, country, class, clientId]
    MonitoringResult:
      type: object
      properties:
        id:
          type: integer
        keywordId:
          type: integer
        keyword:
          type: string
        sourceKeyword:
          type: string
          description: monitoring_keywords.keyword (PHP Keyword column)
        wordToCompare:
          type: string
        appno:
          type: string
        journalDate:
          type: string
        score:
          type: number
        scoreLabel:
          type: string
          description: Very High / High / Medium / Low (PHP MonitoringScoreForHumans)
        conflictClass:
          type: string
        conflictClassDisplay:
          type: string
          description: Pipes replaced with comma-space for display
        conflictCountry:
          type: string
        conflictStatus:
          type: string
        tmAppliedFor:
          type: string
        userDetail:
          type: string
        country:
          type: string
        class:
          type: string
        clientName:
          type: string
        userName:
          type: string
        userEmail:
          type: string
        journalCopyUrl:
          type: string
        favorite:
          type: boolean
          description: PHP TM Watch star / btn-warning when true
      required: [id, keyword, appno]
    MonitoringResultFavoriteToggle:
      type: object
      properties:
        id:
          type: integer
        favorite:
          type: boolean
      required: [id, favorite]
    PaginatedMonitoringResults:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MonitoringResult"
        total:
          type: integer
        page:
          type: integer
        perPage:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, perPage, totalPages]
    ExportBody:
      type: object
      properties:
        type:
          type: string
          enum: [xlsx, pdf]
        dateFrom:
          type: string
        dateTo:
          type: string
        selectedClients:
          type: array
          items:
            type: string
        selectedCountries:
          type: array
          items:
            type: string
        filters:
          type: object
      required: [type]
    ExportResponse:
      type: object
      properties:
        status:
          type: integer
        file:
          type: string
      required: [status]
    ImportResponse:
      type: object
      properties:
        status:
          type: integer
        added:
          type: integer
        message:
          type: string
      required: [status, added]
...